move out the utilities from pipeline_utils.py
